Adds basic requesting of dates
A function is added which returns to a given canteen ID all of their
dates (if there exist any), for those dates is also the information
given whether the canteen is opened on that day or not. Those canteen
date information are stored in a list of CanteenDates!

For this function was also a test added.